requirements/base.txt
Django==2.1
django-ipware==2.1.0
django-model-utils==3.1.2
Pillow==5.2.0
psycopg2-binary==2.7.5
pytz==2018.5
sorl-thumbnail==12.4.1
Unipath==1.1
